requirements.txt
python-telegram-bot==21.*
python-dotenv==1.*
requests==2.*
